Extract the observe package (live acquisition + annotated render)
The last big subsystem extraction. observe is the shared read path behind
snapshot and capture and the Tier-B entrypoint external callers use to
observe a page against a corpus:

- observe.Page(ctx, conn, corpus, opts) -> *Result — extract the component
  tree, apply the corpus (matches, view, components, global names), score
  coverage, and optionally extract live property values. Navigation and the
  browser lifecycle stay with the caller; Page only reads current state.
- observe.Format(w, *Result, FormatOpts) — the annotated section output
  (View, Guide, tree, Warnings, Coverage) previously wired up as
  writeOutput; the section writers move with it. WriteTrace / WriteT2Trace /
  CheckRootComponents / ExtractProperties are exported because the offline
  coverage command and the browser component-query resolver share them.

This empties cmd_snapshot.go from 970 lines (pre-coverage-extraction) to
204: runSnapshot and capture are now thin adapters that navigate, call
observe.Page, and hand the Result to observe.Format. The shared string
helper truncateStr becomes render.TruncateRunes; displayName (only the
annotation-gaps printer uses it) moves next to that printer.

Verified against the burrito example: snapshot (full tree + guide +
coverage), snapshot --coverage, snapshot --json (properties), capture,
browser click by component query (property resolution), and coverage
--trace all match the pre-refactor output. Full suite green.
